requirements.txt
pip
discord.py
discord.py[voice]
youtube_dl
pynacl
python-dotenv
ffmpeg
youtube-search
